Move small independent packages to `pkg/`
