+++ cluster/raft/node_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/robinkb/cascade/cluster"
 	"github.com/robinkb/cascade/cluster/fake"
 	"github.com/robinkb/cascade/cluster/raft/qwal"
-	"github.com/robinkb/cascade/server"
+	"github.com/robinkb/cascade/pkg/server"
 	. "github.com/robinkb/cascade/testing"
 	"go.etcd.io/raft/v3"
 )
