+++ testing/e2e/testnode.go
@@ -8,7 +8,8 @@ import (
 
 	"github.com/robinkb/cascade/cluster/raft"
 	"github.com/robinkb/cascade/cluster/raft/qwal"
-	"github.com/robinkb/cascade/process"
+	"github.com/robinkb/cascade/pkg/process"
+	"github.com/robinkb/cascade/pkg/server"
 	"github.com/robinkb/cascade/registry"
 	registryapi "github.com/robinkb/cascade/registry/api/v2"
 	"github.com/robinkb/cascade/registry/store"
@@ -16,7 +17,6 @@ import (
 	clusterstore "github.com/robinkb/cascade/registry/store/driver/cluster"
 	"github.com/robinkb/cascade/registry/store/driver/inmemory"
 
-	"github.com/robinkb/cascade/server"
 	. "github.com/robinkb/cascade/testing" // nolint: staticcheck
 )
